--- a/package.json
+++ b/package.json
@@ -39,7 +39,6 @@
   "devDependencies": {
     "babel-core": "^5.8.25",
     "babel-eslint": "^4.1.1",
-    "babel-loader": "^5.3.2",
     "babel-relay-plugin": "^0.2.4",
     "del": "^1.2.0",
     "envify": "^3.4.0",