I was testing some fbjs changes and got tired of waiting ~23s for each build so I made it faster.

1. Removed duplicate `modules` task. We already run it for `dist` and `dist:min` (where it gets deduped) so we don't need to run it before then.
  
  *(saves ~5s)*
2. Removed the babel webpack loader. We've already transformed our code. `fbjs` is already transformed. We shouldn't need to transform again. There are some differences (https://gist.github.com/zpao/16e13b4a85010c6fbdff) as a result of this - mostly because we don't reprint, though there are also a couple `'use strict'` statements missing but that's probably ok.
  
  *(saves ~6-7s for each `dist` and `dist:min` but they run in parallel so only see the win once)*

I ran a few times before and after to get some stable numbers with a warm disk cache. I dropped from 23s to 12s.